backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "elasticsearch>=9.2.0",
    "fastapi>=0.120.4",
    "ffmpeg>=1.4",
    "google-adk>=1.17.0",
    "markitdown[all]>=0.1.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "sentencepiece>=0.2.1",
    "transformers>=4.48.0",
    "uvicorn[standard]>=0.30.1",
]
